Fixed memory issue

+++ malloc.h
@@ -2,7 +2,7 @@
 
 typedef struct HTTPC_Malloc {
     void* ptr;
-    int size;
+    size_t size;
     struct HTTPC_Malloc* next;
     struct HTTPC_Malloc* prev;
 } HTTPC_Malloc;
@@ -13,6 +13,6 @@ size_t httpc_allocated_max;
 
 HTTPC_Malloc* httpc_malloced;
 
-void* httpc_malloc(int size);
+void* httpc_malloc(size_t size);
 
 void httpc_free(void* ptr);
